ball part
I let the ball launch from the ground with an initial speed. When it's moving, the gravitational force and the frictional force influence its speed. When the x position exceed the x position of the rope( so far I set the x position to be 300 and it's ok to change), the ball starts to move like a spring with damping. When the user pressed the Mouse and does not release, the ball will drop to the ground.